% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabicl-python.R
\name{install_tabicl_dependencies}
\alias{install_tabicl_dependencies}
\title{Install tabicl}
\usage{
install_tabicl_dependencies(envname = "tabicl", method = "auto", ...)
}
\arguments{
\item{envname}{virtual environment name}

\item{method}{method defaults to "auto"}

\item{...}{additional passes for \code{py_install}}
}
\value{
installs to "tabicl" env by default
}
\description{
Install tabicl
}
