% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabicl-package.R
\docType{package}
\name{tabicl-package}
\alias{tabicl}
\alias{tabicl-package}
\title{tabicl: A Tabular Foundation Model for In-Context Learning on Large Data with reticulate}
\description{
Bindings for TabICL, a tabular foundation model. This package provides R bindings to the python library (soda-inria/tabicl) via reticulate.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@proton.me}

}
\keyword{internal}
