% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabicl-pred.R
\name{tabicl_predict}
\alias{tabicl_predict}
\title{tabicl Predictions}
\usage{
tabicl_predict(fit, data, type = "raw", ...)
}
\arguments{
\item{fit}{a fitted model}

\item{data}{a dataset}

\item{type}{'raw' or 'proba'}

\item{...}{for passing parameters to the predict method}
}
\value{
predictions
}
\description{
tabicl Predictions
}
