% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabicl-fit.R
\name{tabicl_fit}
\alias{tabicl_fit}
\title{Fit tabicl Tabular Foundation Model}
\usage{
tabicl_fit(X, y, problem_type = "classification", ...)
}
\arguments{
\item{X}{training data}

\item{problem_type}{either "classification" or "regression"}

\item{...}{for tabicl additional hyperparameters such as \code{batch_size}}

\item{data}{a dataset}
}
\value{
a tabicl fit
}
\description{
Fit tabicl Tabular Foundation Model
}
